+++ tests/client-tests.js
@@ -165,19 +165,37 @@ describe("Hull", () => {
     it("should allow to pass an array of user aliases", () => {
       const hull = new Hull({ id: "562123b470df84b740000042", secret: "1234", organization: "test" });
 
-      const scoped = hull.asUser({ email: "foo@bar.com", alias: ["namespace:123"] });
+      const scoped = hull.asUser({ email: "foo@bar.com", aliases: ["namespace:123"] });
       const scopedJwtClaims = jwt.decode(scoped.configuration().accessToken, scoped.configuration().secret);
       expect(scopedJwtClaims["io.hull.asUser"])
-        .to.eql({ email: "foo@bar.com", alias: ["namespace:123"] });
+        .to.eql({ email: "foo@bar.com", aliases: ["namespace:123"] });
     });
 
     it("should allow to pass an array of account aliases", () => {
       const hull = new Hull({ id: "562123b470df84b740000042", secret: "1234", organization: "test" });
 
-      const scoped = hull.asAccount({ domain: "bar.com", alias: ["namespace:123"] });
+      const scoped = hull.asAccount({ domain: "bar.com", aliases: ["namespace:123"] });
       const scopedJwtClaims = jwt.decode(scoped.configuration().accessToken, scoped.configuration().secret);
       expect(scopedJwtClaims["io.hull.asAccount"])
-        .to.eql({ domain: "bar.com", alias: ["namespace:123"] });
+        .to.eql({ domain: "bar.com", aliases: ["namespace:123"] });
+    });
+
+    it("should allow to pass an object of user service ids", () => {
+      const hull = new Hull({ id: "562123b470df84b740000042", secret: "1234", organization: "test" });
+
+      const scoped = hull.asUser({ email: "foo@bar.com", service_ids: { service: "abc123" } });
+      const scopedJwtClaims = jwt.decode(scoped.configuration().accessToken, scoped.configuration().secret);
+      expect(scopedJwtClaims["io.hull.asUser"])
+        .to.eql({ email: "foo@bar.com", service_ids: { service: "abc123" } });
+    });
+
+    it("should allow to pass an object of account service ids", () => {
+      const hull = new Hull({ id: "562123b470df84b740000042", secret: "1234", organization: "test" });
+
+      const scoped = hull.asAccount({ domain: "bar.com", service_ids: { service: "abc123" } });
+      const scopedJwtClaims = jwt.decode(scoped.configuration().accessToken, scoped.configuration().secret);
+      expect(scopedJwtClaims["io.hull.asAccount"])
+        .to.eql({ domain: "bar.com", service_ids: { service: "abc123" } });
     });
   });
 });